refactor: class integration
